fix: restructure as proper Python package
- Move code into imf_data_mcp package directory
- Update imports to use relative imports
- Include utils.py in the package
